Fixing load commands function
